moved example getters back into examples, kept flagged column, surrounded create validation with try/catch
